Extract FilterableStatusViewHolder to new file
